requirements.txt
streamlit
Pillow
stegano 
opencv-python-headless


